fix QuickOrderBlock not work StandardPage can not set background color
