+++ src/Foundation/Features/MyOrganization/QuickOrderBlock/Index.cshtml
@@ -1,11 +1,11 @@
 ﻿@using Foundation.Features.MyOrganization.QuickOrderBlock
 
-@model QuickOrderBlock
+@model QuickOrderViewModel
 
-<h3>@Html.PropertyFor(m => m.Title)</h3>
+<h3>@Html.PropertyFor(m => m.CurrentBlock.Title)</h3>
 <div class="row">
     <div class="col-12">
-        @Html.PropertyFor(m => m.MainBody)
+        @Html.PropertyFor(m => m.CurrentBlock.MainBody)
     </div>
 </div>
 
@@ -24,7 +24,7 @@
                         @Html.TextBoxFor(x => x.ProductsList[i].ProductName, new { @class = "form-control square-box", @readonly = "readonly" })
                     </div>
                     <div class="form-group col-xs-12 col-sm-12 col-md-6 col-lg-2">
-                        @Html.TextBoxFor(x => x.ProductsList[i].Sku, new { @class = "form-control square-box", required = "required", @class = "position-relative;" })
+                        @Html.TextBoxFor(x => x.ProductsList[i].Sku, new { @class = "form-control square-box position-relative", required = "required" })
                     </div>
                     <div class="form-group col-xs-12 col-sm-12 col-md-6 col-lg-2">
                         @Html.TextBoxFor(x => x.ProductsList[i].UnitPrice, "{0:0.00}", new { @class = "form-control square-box", @readonly = "readonly" })
