+++ src/Foundation/Features/MyOrganization/QuickOrderBlock/QuickOrderBlock.cs
@@ -1,11 +1,8 @@
 using EPiServer.Core;
 using EPiServer.DataAbstraction;
 using EPiServer.DataAnnotations;
-using EPiServer.Security;
 using Foundation.Features.Shared;
 using Foundation.Infrastructure;
-using Mediachase.Commerce.Security;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
 namespace Foundation.Features.MyOrganization.QuickOrderBlock
@@ -24,23 +21,5 @@ namespace Foundation.Features.MyOrganization.QuickOrderBlock
         [CultureSpecific]
         [Display(Name = "Main body", GroupName = SystemTabNames.Content, Order = 10)]
         public virtual XhtmlString MainBody { get; set; }
-
-        [Ignore]
-        [Display(Name = "Product list", Order = 20)]
-        public List<QuickOrderProductViewModel> ProductsList { get; set; }
-
-        [Ignore]
-        [Display(Name = "Returned messages", Order = 30)]
-        public List<string> ReturnedMessages { get; set; }
-
-        [Ignore]
-        public bool HasOrganization
-        {
-            get
-            {
-                var contact = PrincipalInfo.CurrentPrincipal.GetCustomerContact();
-                return contact?.OwnerId != null;
-            }
-        }
     }
 }
