+++ src/Foundation/Features/StandardPage/_standard-page.scss
@@ -12,6 +12,7 @@
     }
 
     &__background {
+        position: absolute;
         left: 0;
         top: 0;
         height: 100%;
